Improve CSSCompiler accuracy and validation

+++ index.js
@@ -1,5 +1,5 @@
 let fs = require('fs');
-let { CSSCompiler, CSSStyle } = require('./compiler');
+let { CSSCompiler, CSSStyle, CSSStyleIntegrityError } = require('./compiler');
 
 console.log('Preparing files...');
 
@@ -14,18 +14,18 @@ function fetchUserStyleProperties(contents) {
     let begin = contents.indexOf(start);
     let finish = contents.indexOf(end);
 
+    if(begin === -1 || finish === -1) {
+        return '';
+    }
+
     let properties = '';
     for(let i = begin; i < finish + end.length; i++) {
         properties += contents[i];
     }
-    return {
-        properties: properties,
-        lastIndex: finish + end.length
-    }
+    return properties;
 }
 
-let { properties, lastIndex } = fetchUserStyleProperties(contents);
-let remaining = contents.substring(lastIndex, contents.length);
+let properties = fetchUserStyleProperties(contents);
 
 console.log('Preparing compiler...');
 
@@ -34,27 +34,47 @@ compiler.init();
 
 console.log('Reading...');
 
-const uncompressed = compiler.purge(remaining);
+const uncompressed = compiler.interpret(contents);
 const compressed = new CSSStyle();
 
 console.log('Merging...');
 compiler.mergeStyles(compressed, uncompressed.styles);
 
-const uncompressedResult = compiler.toString(uncompressed, {
-    textIndent: true
-});
-const compressedResult = compiler.toString(compressed, {
-    textIndent: true
-});
+// DEBUG START
+// compressed.styles[20].type = 2; // Changing CSS style type into media-query
+//
+// compressed.styles[10].styles.splice(26, 59); // Remove a selector from compressed stylesheet
+//
+// compressed.styles[47].documentText = '@-moz-document url-prefix("https://fake.naver.com")'; // Wrong document selector
+// DEBUG END
 
-console.log(`Uncompressed size: ${uncompressedResult.length}`);
-console.log(`Compressed size: ${compressedResult.length}`);
-console.log(`Total compressed: -${(100 - (compressedResult.length * 100 / uncompressedResult.length)).toFixed(1)}%`);
+// Verify uncompressed and compressed styles integrity.
+const errors = compiler.verifyStyleIntegrity(uncompressed, compressed);
+if(errors.length > 0) {
+    console.log('Original styles and compressed styles\' components are not identical.');
+    console.log('Caused by:');
+    for(let error of errors) {
+        console.log(error.toString());
+    }
+} else {
+    console.log('Original styles and compressed styles\' components are identical.');
 
-console.log('Writing them in file...');
+    const uncompressedResult = compiler.toString(uncompressed, {
+        textIndent: true
+    });
+    const compressedResult = compiler.toString(compressed, {
+        textIndent: true
+    });
 
-if(fs.existsSync(compressedStylePath)) {
-    fs.unlinkSync(compressedStylePath);
+    console.log(`Uncompressed size: ${uncompressedResult.length}`);
+    console.log(`Compressed size: ${compressedResult.length}`);
+    console.log(`Total compressed: -${(100 - (compressedResult.length * 100 / uncompressedResult.length)).toFixed(1)}%`);
+
+    console.log('Writing them in file...');
+
+    if(fs.existsSync(compressedStylePath)) {
+        fs.unlinkSync(compressedStylePath);
+    }
+
+    fs.appendFileSync(compressedStylePath, properties + '\n' + compressedResult, 'utf8');
 }
-
-fs.appendFileSync(compressedStylePath, properties + '\n' + compressedResult, 'utf8');
